backend/requirements.txt
fastapi[standard]==0.115.8
haystack-ai==2.9.0
python-dotenv==1.0.1
sentence-transformers==3.4.1